perf: use native package extensions in julia 1.9+
