+++ Project.toml
@@ -13,17 +13,25 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
+[weakdeps]
+Catalyst = "479239e8-5488-4da2-87a7-35f2df7eef83"
+ModelingToolkit = "961ee093-0014-501f-94e3-6117800e7a78"
+Petri = "4259d249-1051-49fa-8328-3f8ab9391c33"
+Requires = "ae029012-a4dd-5104-9daa-d747884805df"        # makes sure Requires isn't a dependency for Julia v1.9+
+
+[extensions]
+AlgebraicPetriPetriExt = "Petri"
+AlgebraicPetriCatalystExt = "Catalyst"
+AlgebraicPetriModelingToolkitExt = "ModelingToolkit"
+
 [compat]
 Catalyst = "13"
 Catlab = "0.14.16"
 DataStructures = "0.18"
 GeneralizedGenerated = "0.3"
 LabelledArrays = "1"
+ModelingToolkit = "8"
+Petri = "1"
 Requires = "1"
 StatsBase = "0.33,0.34"
 julia = "1.6"
-
-[extras]
-Catalyst = "479239e8-5488-4da2-87a7-35f2df7eef83"
-Petri = "4259d249-1051-49fa-8328-3f8ab9391c33"
-Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
