+++ src/interoperability.jl
@@ -1,31 +1,7 @@
 using Requires
 
 function __init__()
-  @require Petri="4259d249-1051-49fa-8328-3f8ab9391c33" begin
-    import .Petri
-
-    # Interoperability with Petri.jl
-    Petri.Model(p::AbstractPetriNet) = begin
-      ts = TransitionMatrices(p)
-
-      if has_subpart(p, :sname) && has_subpart(p, :tname)
-        snames = [sname(p, s) for s in 1:ns(p)]
-        tnames = [tname(p, t) for t in 1:nt(p)]
-        t_in = map(i->LVector(;[(snames[k]=>v) for (k,v) in enumerate(ts.input[i,:]) if v != 0]...), 1:nt(p))
-        t_out = map(i->LVector(;[(snames[k]=>v) for (k,v) in enumerate(ts.output[i,:]) if v != 0]...), 1:nt(p))
-        Δ = LVector(;[(tnames[i]=>t) for (i,t) in enumerate(zip(t_in, t_out))]...)
-        S = collect(values(snames)) 
-      else
-        t_in = map(i->Dict(k=>v for (k,v) in enumerate(ts.input[i,:]) if v != 0), 1:nt(p))
-        t_out = map(i->Dict(k=>v for (k,v) in enumerate(ts.output[i,:]) if v != 0), 1:nt(p))
-        Δ = Dict(i=>t for (i,t) in enumerate(zip(t_in, t_out)))
-        S = ns(p)
-      end
-
-      return Petri.Model(ns(p), Δ)
-    end
-  end
-
-  @require Catalyst="479239e8-5488-4da2-87a7-35f2df7eef83" include("CatalystInterop.jl")
-  @require ModelingToolkit="961ee093-0014-501f-94e3-6117800e7a78" include("ModelingToolkitInterop.jl")
+  @require Petri="4259d249-1051-49fa-8328-3f8ab9391c33" include("../ext/AlgebraicPetriPetriExt.jl")
+  @require Catalyst="479239e8-5488-4da2-87a7-35f2df7eef83" include("../ext/AlgebraicPetriCatalystExt.jl")
+  @require ModelingToolkit="961ee093-0014-501f-94e3-6117800e7a78" include("../ext/AlgebraicPetriModelingToolkitExt.jl")
 end
